% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obiscrapper.R
\name{connectobi}
\alias{connectobi}
\title{connectobi}
\usage{
connectobi(
  path_to_firefox = NA,
  username = NA,
  password = NA,
  obilink = NA,
  extranet_elem_id = NA,
  debuglevel = 0
)
}
\arguments{
\item{path_to_firefox}{Optional path to Firefox in case it is not available in your path. Default NA.}

\item{username}{Username to access the OBI server. If not provided, it will be prompted.}

\item{password}{User password to access the OBI server. If not provided, it will be prompted.}

\item{obilink}{Mandatory link to the analytics web portal. Eg https://mycompany.int/analytics/".}

\item{extranet_elem_id}{Optional. Id of the HTML element for user and password, in case you use an extranet before analytics server. Example: list("user"="rad_usr", "password"="rad_pw")}
}
\description{
Open Firefox, browse the OBI web portal and provide your username and password to the formular so you are logged in.
}
\examples{
# Connect without providing the password to the function. Password will be prompted and hidden (using package getPass)
connectobi(username="myusername", obilink="https://mycompany.int/analytics/")

# Connect and provide the password to the function. Nothing will be prompted.
# Useful for batch mode.
connectobi(username="myusername", password="myfancypassword", obilink="https://mycompany.int/analytics/")
}
