% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obiscraper.R
\name{disconnectobi}
\alias{disconnectobi}
\title{disconnectobi}
\usage{
disconnectobi()
}
\value{
None.
}
\description{
Close the session, RSelenium server, and Firefox browser
}
\examples{
disconnectobi()
}
