% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obiscraper.R
\name{get_rs_client}
\alias{get_rs_client}
\title{get_rs_client}
\usage{
get_rs_client()
}
\value{
RSelenium::rsDriver[['client']]
}
\description{
Return the RSelenium client, for debugging
}
\examples{
get_rs_client()
}
