% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obiscrapper.R
\docType{package}
\name{obiscrapper}
\alias{obiscrapper}
\title{obiscrapper: A Tool For Submitting Queries to Oracle Business Intelligence by webscrapping the analytics server}
\description{
Execute OBI queries by web scrapping the analytics server with RSelenium. The analytics login form is handled,
even when behind an extranet. Then queries are submitted by using
the OBI GO URL (saw.dll?Go&SQL=select+Region,Euro+from+SupplierSalesEurope).
Queries result returned as R Data Frame.
}
\details{
Prerequisite:\cr
-CRAN package RSelenium\cr
-Firefox

Bugs report:\cr
 \url{https://github.com/sthonnard/obiscrapper}
}
\section{obiscrapper functions}{

\strong{connectobi(path_to_firefox, username, password, obilink)}\cr
Open Firefox, browse the OBI web portal and provide your username in passward to the formular so you are logged in.

\strong{query_obi(query)}\cr
Submit a query to the OBI web portal by using the GO URL and return the result as a Data Frame.\cr

\strong{disconnectobi()}\cr
Disconnect from OBI and close the web browser.\cr

\strong{get_obi_client()}\cr
Return the RSelenium driver client, for debugging.
}

