% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obiscraper.R
\name{submit_physical_sql}
\alias{submit_physical_sql}
\title{submit_physical_sql}
\usage{
submit_physical_sql(
  psql = "select distinct icao_code from dwh.airport",
  pconnectionpool = "my_dwh",
  clean_final_result = TRUE
)
}
\arguments{
\item{psql}{SQL statement that will be submitted to connection pool pconnectionpool of the OBI physical layer.}

\item{pconnectionpool}{Name of the connetion pool.}

\item{clean_final_result}{Clean final result (TRUE/FALSE). Will removed ",00" from integers.}
}
\value{
Data Frame containing the query result.
}
\description{
Sumbit SQL statement to the OBI physical layer and return the result as a Data Frame. Experimental!
}
\examples{
run_physical_sql(psql="select distinct icao_code from dwh.airport", pconnectionpool="my_dwh")
}
