% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obiscrapper.R
\name{submit_query}
\alias{submit_query}
\title{submit_query}
\usage{
submit_query(
 
    query = "SELECT ... FROM \\"...\\" ORDER BY 1 ASC NULLS LAST FETCH FIRST 10000000 ROWS ONLY"
)
}
\arguments{
\item{query}{The query that will be sent to OBI}
}
\description{
Submit a query to OBI and get the result as a dataframe
}
\examples{
submit_query('SELECT "MyModel"."Flights"."Departure Airport" s_1
FROM "MyModel"
ORDER BY 1 ASC NULLS LAST
FETCH FIRST 10000000 ROWS ONL')
}
