% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obiscraper.R
\name{submit_query}
\alias{submit_query}
\title{submit_query}
\usage{
submit_query(
 
    query = "SELECT ... FROM \\"...\\" ORDER BY 1 ASC NULLS LAST FETCH FIRST 10000000 ROWS ONLY"
)
}
\arguments{
\item{query}{The query that will be sent to Oracle Business Intelligence Presentation Layer.}
}
\value{
Data Frame containing the query result.
}
\description{
Submit a query to the Oracle Business Intelligence Presentaion Layer and get the result as a Data Frame.
}
\details{
This procedure will use the Oracle Business Intelligence GO URL facility (eg saw.dll?Go&SQL=select+Region,Euro+from+SupplierSalesEurope)
in order to download the result as a flat file in your temp folder, which will then be parsed as a Data Frame in your R session.
}
\examples{
# submit_query('SELECT "MyModel"."Flights"."Departure Airport" s_1 FROM "MyModel" ORDER BY 1 ASC NULLS LAST FETCH FIRST 10000000 ROWS ONL')
}
